Fold ChannelMetadataSerializer into DeviceChannelMetadataSerializer
Its only consumer was the device plugin's viewset, a plain
ReadOnlyModelViewSet that needs the model-instance-based fields.
